+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2026-09-26 06:34]
+
+### Added
+- Template images: add images to a template using `[[NAME]]` syntax, same flow as variables (add first, then use in content)
+- Searchable image picker with thumbnail preview when adding an image to a template
+
 ## [2026-09-26 03:50]
 
 ### Added